% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{addFlowJoGatesInfo}
\alias{addFlowJoGatesInfo}
\title{add FlowJo manual gates membership information to a flow frame}
\usage{
addFlowJoGatesInfo(ff, gateNames, wspFile, ...)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{gateNames}{a character vector, containing the names of the
flowJo gates to be checked}

\item{wspFile}{a flowjo workspace}

\item{...}{Extra arguments passed to \code{getFlowJoLabels()}}
}
\value{
the input flowFrame, filled with additional columns of logical
one per checked gate. The name of the new columns match the name of the
gates.
}
\description{
reads a flowjo workspace file using the CytoML package,
and extract cell by cell information as to whether the considered cell is
within or outside user chosen gates.
Note that not only the selected gate, but also
all its hierarchy of gates, will be checked when extracting the gate
membership information.
}
