% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{anonymizeMarkers}
\alias{anonymizeMarkers}
\title{anonymizing some markers of a flowFrame}
\usage{
anonymizeMarkers(
  ff,
  oldMarkerNames,
  newMarkerNames,
  newExperimentName = NULL,
  ...
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{oldMarkerNames}{the marker names to be amended, or alternatively,
the channel names for which to update the marker names}

\item{newMarkerNames}{the new marker names to be given
to the provided \code{oldMarkerNames}}

\item{newExperimentName}{(optional) a new experiment name to be given to
the flowFrame}

\item{...}{other arguments (not used)}
}
\value{
a new flowCore::flowFrame with the updated marker names
(and possibly, the new experiment name)
}
\description{
: in a flowCore::flowFrame, update some of the marker names.
This translated into :
\itemize{
\item updating the desc field of the parameters pheno data dataframe
\item updating the corresponding keyword value in the flowFrame.
}
}
\examples{

data(OMIP021Samples)

retFF <- anonymizeMarkers(OMIP021Samples[[1]],
                          oldMarkerNames = c("FSC-A","BV785 - CD3"),
                          newMarkerName = c("Fwd Scatter-A", "CD3"),
                          newExperimentName = "MyExperiment")

}
