% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{applyFlowJoGate}
\alias{applyFlowJoGate}
\title{perform manual gating from a FlowJo gate}
\usage{
applyFlowJoGate(ff, wspFile, gateName, ...)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{wspFile}{a flowjo workspace}

\item{gateName}{the name of the flowJo gate that will be applied}

\item{...}{Extra arguments passed to \code{getFlowJoLabels()}}
}
\value{
a list with :
\itemize{
\item the first element ("matrix") is a matrix containing filtering results for
each specified gate
\item the second element ("labels") is a vector which assigns one label to each
cell. If no cell type correspond to a specific cell, than the keyword
'unlabeled' is assigned to this cell. If the cell belongs to several gates
(meaning that the gates ar not disjoints), than this cell is assigned to the
gate with the less matching cells.
}
}
\description{
perform manual gating by: reading a flowjo workspace file
using the flowWorkspace library, and apply one of the gates
to the inputflowFrame. Note that not only the selected gate, but also
all its hierarchy of gates, will be applied to the input dataset.
}
