% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoPipelineUtils-package.R
\docType{package}
\name{CytoPipelineUtils-package}
\alias{CytoPipelineUtils}
\alias{CytoPipelineUtils-package}
\title{CytoPipelineUtils: Utils for use with CytoPipeline}
\description{
This package is meant to be used in conjunction of CytoPipeline package, which provides support for automation and visualization of flow cytometry data analysis pipelines. CytoPipelineUtils provides a series of wrapper implementations that can in turn be defined as CytoProcessingSteps in CytoPipeline. It is therefore a helper package, which is aimed at hosting wrapper implementations of various functions of various packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://uclouvain-cbio.github.io/CytoPipeline}
  \item Report bugs at \url{https://github.com/UCLouvain-CBIO/CytoPipelineUtils/issues}
}

}
\author{
\strong{Maintainer}: Philippe Hauchamps \email{philippe.hauchamps@uclouvain.be} (\href{https://orcid.org/0000-0003-2865-1852}{ORCID})

Authors:
\itemize{
  \item Laurent Gatto \email{laurent.gatto@uclouvain.be} (\href{https://orcid.org/0000-0002-1520-2268}{ORCID})
}

Other contributors:
\itemize{
  \item Dan Lin \email{dan.8.lin@gsk.com} [contributor]
}

}
\keyword{internal}
