% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{getEventNbFromFJGates}
\alias{getEventNbFromFJGates}
\title{perform manual gating from a FlowJo gate}
\usage{
getEventNbFromFJGates(
  ff,
  wspFile,
  gates = NULL,
  defaultGates = c("leaves", "all"),
  ...
)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{wspFile}{a flowjo workspace}

\item{gates}{vector of flowJo gate nodes to parse (should be common to
all groups!). Default NULL then uses either all leaf node from first group,
or all gates from first group, depending on 'defaultGates'.}

\item{defaultGates}{if gates are not specified, than triggers the
choice of either :
\itemize{
\item all leaf nodes ('leaves') (the default)
\item all nodes ('all')
}}

\item{...}{Extra arguments passed to \code{getFlowJoLabels()}}
}
\value{
a dataframe with the nb of events per gate, including one additional
'unlabelled' gate for un-labelled events.
}
\description{
perform manual gating by: reading a flowjo workspace file
using the flowWorkspace library, and apply the selected gates to the input
flowFrame. Then provides nb of remaining events for the selected gates.
}
