% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{removeDebrisFlowClust}
\alias{removeDebrisFlowClust}
\title{remove debris from a flowFrame, using flowClust}
\usage{
removeDebrisFlowClust(ff, FSCChannel, SSCChannel, nClust, probaLevel, ...)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{FSCChannel}{the name of the FSC channel}

\item{SSCChannel}{the name of the SSC channel}

\item{nClust}{number of clusters to identify}

\item{probaLevel}{the probability level}

\item{...}{additional parameters passed to flowClust::flowClust()}
}
\value{
a flowCore::flowFrame with removed debris events from the input
}
\description{
this function removes debris from a flowFrame,
using clustering capabilities of flowClust::tmixFilter(). The idea is to
pre-select a number of clusters to be found in the (FSC,SSC) 2D view, and
select all clusters behalve the one closest nearest the origin.
Then we take all events that are inside the \code{probaLevel} quantile curve
for at least one of the remaining clusters.
}
\examples{

rawDataDir <- system.file("extdata", package = "CytoPipeline")
sampleFiles <-
    file.path(rawDataDir, list.files(rawDataDir, pattern = "Donor"))

truncateMaxRange <- FALSE
minLimit <- NULL

# create flowCore::flowSet with all samples of a dataset
fsRaw <- readSampleFiles(
    sampleFiles = sampleFiles,
    whichSamples = "all",
    truncate_max_range = truncateMaxRange,
    min.limit = minLimit)

suppressWarnings(ff_m <- removeMarginsPeacoQC(x = fsRaw[[2]]))
    
ff_c <-
    compensateFromMatrix(ff_m,
                         matrixSource = "fcs")        


ff_cells <-
    removeDebrisFlowClust(ff_c,
                          FSCChannel = "FSC-A",
                          SSCChannel = "SSC-A",
                          nClust = 3,
                          probaLevel = 0.97,
                          B = 100)

}
