% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{removeMarginsFlowAI}
\alias{removeMarginsFlowAI}
\title{remove margin events, using flowAI}
\usage{
removeMarginsFlowAI(x, ...)
}
\arguments{
\item{x}{a flowCore::flowFrame or a flowCore::flowSet}

\item{...}{additional parameters passed to flowAI::flow_auto_qc(), apart
from the following ones : remove_from, output, ChExcludeFM, html_report,
mini_report, fcs_QC, fcs_highQ, fcs_lowQ, folder_results}
}
\value{
a flowCore::flowFrame with removed doublets events from the input
}
\description{
wrapper around flowAI::flow_auto_qc(), with inputs directed
towards specifically remove the margin events.
In the current implementation, all the signal channels, i.e. both scatter
and fluo channels are scanned.
}
\examples{

rawDataDir <- system.file("extdata", package = "CytoPipeline")
sampleFiles <-
    file.path(rawDataDir, list.files(rawDataDir, pattern = "Donor"))

truncateMaxRange <- FALSE
minLimit <- NULL

# create flowCore::flowSet with all samples of a dataset
fsRaw <- readSampleFiles(
    sampleFiles = sampleFiles,
    whichSamples = "all",
    truncate_max_range = truncateMaxRange,
    min.limit = minLimit)

ff_m <- removeMarginsFlowAI(x = fsRaw[[2]])
    
}
