% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoProcessingStepImplementations.R
\name{applyFlowJoGate}
\alias{applyFlowJoGate}
\title{perform manual gating from a FlowJo gate}
\usage{
applyFlowJoGate(ff, gateName, wspFile, ...)
}
\arguments{
\item{ff}{a flowCore::flowFrame}

\item{gateName}{the name of the flowJo gate that will be applied}

\item{wspFile}{a flowjo workspace}

\item{...}{Extra arguments passed to \code{getFlowJoLabels()}}
}
\value{
a flowFrame containing only the events inside the gate
}
\description{
perform manual gating by: reading a flowjo workspace file
using the CytoML package, and apply one of the gates
to the input flowFrame. Note that not only the selected gate, but also
all its hierarchy of gates, will be applied to the input dataset.
}
